+++ src/apps/DisciplineReleaseControl/Components/Sidesheet/ReleaseControlSidesheet.tsx
@@ -20,7 +20,6 @@ import { SidesheetTabList } from './SidesheetTabs';
 import { TablesTab, WarningBanner, WarningBannerText } from './styles';
 import { WorkOrderTab } from './WorkOrderTab';
 
-
 interface GatewaySidesheetProps {
     item: Pipetest | HTSidesheet;
     actions: SidesheetApi;
@@ -79,71 +78,76 @@ export const ReleaseControlSidesheet = ({
 
     return (
         <Wrapper>
-            <ModelViewerContextProvider>
-                <ReleaseControlErrorBanner message={errorMessage} />
-                <ReleaseControlSidesheetBanner pipetest={item} />
-                <Tabs activeTab={activeTab} onChange={handleChange}>
-                    <SidesheetTabList>
-                        <Tabs.Tab>Circuit diagram</Tabs.Tab>
-                        <Tabs.Tab>Work orders</Tabs.Tab>
-                        <Tabs.Tab>Insulation</Tabs.Tab>
-                        <Tabs.Tab>Checklists</Tabs.Tab>
-                        {!isProduction() && <Tabs.Tab>3D-visualisation</Tabs.Tab>}
-                    </SidesheetTabList>
-                    <Tabs.Panels>
-                        <Tabs.Panel>
-                            <ElectroView
-                                pipetest={item}
-                                pipetests={data !== undefined ? data : []}
-                                width={width}
-                            />
-                        </Tabs.Panel>
-                        <Tabs.Panel>
-                            <WorkOrderTab id={item.name} />
-                        </Tabs.Panel>
+            <ReleaseControlErrorBanner message={errorMessage} />
+            <ReleaseControlSidesheetBanner pipetest={item} />
+            <Tabs activeTab={activeTab} onChange={handleChange}>
+                <SidesheetTabList>
+                    <Tabs.Tab>Circuit diagram</Tabs.Tab>
+                    <Tabs.Tab>Work orders</Tabs.Tab>
+                    <Tabs.Tab>Insulation</Tabs.Tab>
+                    <Tabs.Tab>Checklists</Tabs.Tab>
+                    {!isProduction() && <Tabs.Tab>3D-visualisation</Tabs.Tab>}
+                </SidesheetTabList>
+                <Tabs.Panels>
+                    <Tabs.Panel>
+                        <ElectroView
+                            pipetest={item}
+                            pipetests={data !== undefined ? data : []}
+                            width={width}
+                        />
+                    </Tabs.Panel>
+                    <Tabs.Panel>
+                        <WorkOrderTab id={item.name} />
+                    </Tabs.Panel>
 
-                        <Tabs.Panel>
-                            {missingInsulationCheckListsCount !== 0 &&
-                                (missingInsulationCheckListsCount === 1 ? (
-                                    <WarningBanner>
-                                        <WarningBannerText>
-                                            ! Warning: {missingInsulationCheckListsCount} insulation
-                                            box missing checklists in ProCoSys.
-                                        </WarningBannerText>
-                                    </WarningBanner>
-                                ) : (
-                                    <WarningBanner>
-                                        <WarningBannerText>
-                                            ! Warning: {missingInsulationCheckListsCount} insulation
-                                            boxes missing checklists in ProCoSys.
-                                        </WarningBannerText>
-                                    </WarningBanner>
-                                ))}
-                            <TablesTab>
-                                <InsulationTable
-                                    insulations={item.pipeInsulationBoxes}
-                                    pipeInsulation={true}
-                                />
-                                <br />
-                                <InsulationTable
-                                    insulations={item.insulationBoxes}
-                                    pipeInsulation={false}
-                                    pipetestName={item.name}
-                                />
-                            </TablesTab>
-                        </Tabs.Panel>
-                        <Tabs.Panel>
-                            <TablesTab>
-                                <h4>{item.description}</h4>
-                                <CheckListTable checkLists={item.checkLists} />
-                            </TablesTab>
-                        </Tabs.Panel>
-                        {!isProduction() && (
-                            <Panel>{activeTab === 4 && <ThreeDView pipetest={item} />}</Panel>
-                        )}
-                    </Tabs.Panels>
-                </Tabs>
-            </ModelViewerContextProvider>
+                    <Tabs.Panel>
+                        {missingInsulationCheckListsCount !== 0 &&
+                            (missingInsulationCheckListsCount === 1 ? (
+                                <WarningBanner>
+                                    <WarningBannerText>
+                                        ! Warning: {missingInsulationCheckListsCount} insulation box
+                                        missing checklists in ProCoSys.
+                                    </WarningBannerText>
+                                </WarningBanner>
+                            ) : (
+                                <WarningBanner>
+                                    <WarningBannerText>
+                                        ! Warning: {missingInsulationCheckListsCount} insulation
+                                        boxes missing checklists in ProCoSys.
+                                    </WarningBannerText>
+                                </WarningBanner>
+                            ))}
+                        <TablesTab>
+                            <InsulationTable
+                                insulations={item.pipeInsulationBoxes}
+                                pipeInsulation={true}
+                            />
+                            <br />
+                            <InsulationTable
+                                insulations={item.insulationBoxes}
+                                pipeInsulation={false}
+                                pipetestName={item.name}
+                            />
+                        </TablesTab>
+                    </Tabs.Panel>
+                    <Tabs.Panel>
+                        <TablesTab>
+                            <h4>{item.description}</h4>
+                            <CheckListTable checkLists={item.checkLists} />
+                        </TablesTab>
+                    </Tabs.Panel>
+                    {!isProduction() && (
+                        <Panel>
+                            {activeTab === 4 && (
+                                <ModelViewerContextProvider>
+                                    <ThreeDView pipetest={item} />
+                                </ModelViewerContextProvider>
+                            )}
+                        </Panel>
+                    )}
+                </Tabs.Panels>
+            </Tabs>
+            \
         </Wrapper>
     );
 };
